requirements.txt
python=3.6
pip
tensorflow==1.7
Pillow>=4.2.1
matplotlib
numpy>=1.11.0
jupyter
pytest>=3.2.2
docopt
pyyaml
protobuf==3.5.2
grpcio==1.11.0
Pytorch==1.*
pandas
scipy
ipykernel
unityagents==0.4.0
commentjson
torchsummary